tests/wpt/web-platform-tests/app-history/navigate/reload-rejection-order-unload-unserializablestate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  let navigate_event_count = 0;
  i.contentWindow.appHistory.onnavigate = () => navigate_event_count++;
  i.contentWindow.appHistory.onnavigatesuccess = t.unreached_func('onnavigatesuccess should not be called');
  i.contentWindow.appHistory.onnavigateerror = t.unreached_func('onnavigateerror should not be called');
  i.contentWindow.appHistory.navigate("?1");
  await new Promise(resolve => {
    i.contentWindow.onunload = t.step_func(() => {
      i.contentWindow.appHistory.reload({ state: document.body }).catch(t.step_func(e => {
        assert_equals(e.constructor, i.contentWindow.DOMException);
        assert_equals(e.name, "InvalidStateError");
        resolve();
      }));
    });
  });
  assert_equals(navigate_event_count, 1);
}, `reload() with an unserializable state inside onunload throws "InvalidStateError", not "DataCloneError"`);
</script>
